apps/servers/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
sqlalchemy==2.0.30
alembic==1.13.1
python-jose==3.3.0
passlib==1.7.4
pydantic==2.6.4
pytest==8.2.0
httpx==0.27.0
black==24.4.2
isort==5.13.2
